+++ src/com/intellij/idea/plugin/hybris/acl/formatting/AclCodeStyleSettingsProvider.kt
@@ -1,0 +1,42 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.intellij.idea.plugin.hybris.acl.formatting
+
+import com.intellij.application.options.CodeStyleAbstractConfigurable
+import com.intellij.application.options.TabbedLanguageCodeStylePanel
+import com.intellij.idea.plugin.hybris.acl.AclLanguage
+import com.intellij.idea.plugin.hybris.common.HybrisConstants
+import com.intellij.psi.codeStyle.CodeStyleSettings
+import com.intellij.psi.codeStyle.CodeStyleSettingsProvider
+
+class AclCodeStyleSettingsProvider : CodeStyleSettingsProvider() {
+    override fun createCustomSettings(settings: CodeStyleSettings) = AclCodeStyleSettings(settings)
+
+    override fun getConfigurableDisplayName() = HybrisConstants.ACL
+
+    override fun createConfigurable(settings: CodeStyleSettings, originalSettings: CodeStyleSettings) = object : CodeStyleAbstractConfigurable(
+        settings,
+        originalSettings,
+        HybrisConstants.ACL
+    ) {
+        override fun createPanel(settings: CodeStyleSettings) = SimpleCodeStyleMainPanel(currentSettings, settings)
+        override fun getHelpTopic() = null
+    }
+
+    private class SimpleCodeStyleMainPanel(currentSettings: CodeStyleSettings, settings: CodeStyleSettings) : TabbedLanguageCodeStylePanel(AclLanguage, currentSettings, settings)
+}